# openapi.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Dushuli
  description: "读书历CRUD服务"
  license:
    name: MIT
servers:
  - url: https://api.36node.com{basePath}
    description: The production API server
    variables:
      basePath:
        default: /dushuli/v0
paths:
...
  /members:
    get:
      summary: list all members
      operationId: listMembers
      tags:
        - member
      parameters:
        - name: _select
          in: query
          description: select which fields from db
          required: false
          schema:
            type: string
        - name: _limit
          in: query
          description: how many items to return at one time (max 100)
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: _offset
          in: query
          description: How many items will be skip
          required: false
          schema:
            type: string
        - name: _sort
          in: query
          description: specifies which document fields to sort
          required: false
          schema:
            type: string
        - name: users
          in: query
          required: false
          schema:
            type: string
        - name: active
          in: query
          required: false
          schema: 
            type: boolean
      responses:
        "200":
          description: a paged array of members
          headers:
            X-Total-Count:
              description: total count of all members match the query
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
    
    post:
      summary: create a member
      operationId: createMember
      tags:
        - member
      requestBody:
        description: member to be created
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Member" 
      responses:
        "201":
          description: the member created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /members/{user}:
    get:
      summary: get member by user
      operationId: getMember
      tags:
        - member
      parameters:
        - name: user
          in: path
          required: true
          description: the user of the member to retrieve
          schema:
            type: string
      responses:
        "200":
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
        
    delete:
      summary: delete a member
      operationId: deleteMember
      tags:
        - member
      parameters:
        - name: user
          in: path
          description: the user of the member to be deleted
          required: true
          schema:
            type: string
      responses:
        "204":
          description: member deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      summary: update a member
      operationId: updateMember
      tags:
        - member
      parameters:
        - name: user
          in: path
          description: the user of the member to be updated
          required: true
          schema:
            type: string
      requestBody:
        description: member to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Member"
      responses:
        "200":
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"    
...
  /orders:
    get:
      summary: list all orders
      operationId: listOrders
      tags:
        - order
      parameters:
        - name: _select
          in: query
          description: select which fields from db
          required: false
          schema:
            type: string
        - name: _limit
          in: query
          description: how many items to return at one time (max 100)
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: _offset
          in: query
          description: How many items will be skip
          required: false
          schema:
            type: string
        - name: _sort
          in: query
          description: specifies which document fields to sort
          required: false
          schema:
            type: string
        - name: paid
          in: query
          required: false
          schema:
            type: boolean
        - name: method
          in: query
          required: false
          schema:
            type: string
        - name: createdBy
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: a paged array of orders
          headers:
            X-Total-Count:
              description: total count of all orders match the query
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
    
    post:
      summary: create a order
      operationId: createOrder
      tags:
        - order
      requestBody:
        description: order to be created
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Order" 
      responses:
        "201":
          description: the order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /orders/{orderId}:
    get:
      summary: get order by orderId
      operationId: getOrder
      tags:
        - order
      parameters:
        - name: orderId
          in: path
          required: true
          description: the orderId of the order to retrieve
          schema:
            type: string
      responses:
        "200":
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
        
    delete:
      summary: delete a order
      operationId: deleteOrder
      tags:
        - order
      parameters:
        - name: orderId
          in: path
          description: the orderId of the order to be deleted
          required: true
          schema:
            type: string
      responses:
        "204":
          description: order deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      summary: update order
      operationId: updateOrder
      tags:
        - order
      parameters:
        - name: orderId
          in: path
          description: the user of the order to be updated
          required: true
          schema:
            type: string
      requestBody:
        description: order to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Order"
      responses:
        "200":
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"    
security:
  - bearerAuth: []
components:
  responses:
...
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Err"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Err"
    InternalError:
      description: unexpected error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Err"
  schemas:
...
    Period:
      type: object
      properties:
        start: 
          type: string
          format: date-time
          description: 有效期起始时间
        end:
          type: string
          format: date-time
          description: 有效期结束时间
        trial:
          type: boolean
          description: 是否是试用
        product:
          type: string
          description: 商品 id
        active: 
          type: string
          description: 是否有效
    Member:
      type: object
      required:
        - user
      properties:
        id:
          type: string
          description: id
        createdAt:
          type: string
          format: date-time
          description: 创建时间
        updatedAt:
          type: string
          format: date-time
          description: 更新时间
        user:
          type: string
          description: 用户id
        period:
          type: array
          items:
            $ref: "#/components/schemas/Period"
        active:
          type: boolean
          description: 是否激活
    Order:
      type: object
      properties:
        id:
          type: string
          description: id
        createdAt:
          type: string
          format: date-time
          description: 创建时间
        updatedAt:
          type: string
          format: date-time
          description: 更新时间
        no:
          type: string
          description: 订单号
        paid:
          type: boolean
          description: 是否付款
        product:
          type: string
          description: 商品 id
        method:
          type: string
          description: 支付方式
        paidAt:
          type: string
          format: date-time
          description: 付款时间
        data:
          type: object
          description: 支付回调数据
        createdBy:
          type: string
          description: 购买的用户 id
...
    Err:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT